refactor!: Drop posthog integration
We don't use this anymore, feel free to copy and build your plugin/app
if you want.
