+++ .eslintrc
@@ -43,7 +43,6 @@
 		"is_null": true,
 		"in_list": true,
 		"has_common": true,
-		"posthog": true,
 		"has_words": true,
 		"validate_email": true,
 		"open_web_template_values_editor": true,
