+++ frappe/public/js/telemetry/index.js
@@ -1,10 +1,8 @@
-import { posthog_provider } from "./posthog.js";
 import { pulse_provider } from "./pulse.js";
 
 class TelemetryManager {
 	constructor() {
 		this.enabled = frappe.boot.enable_telemetry || false;
-		this.posthog_available = Boolean(frappe.boot.telemetry_provider?.includes("posthog"));
 		this.pulse_available = Boolean(frappe.boot.telemetry_provider?.includes("pulse"));
 
 		this.init_providers();
@@ -12,12 +10,6 @@ class TelemetryManager {
 
 	init_providers() {
 		this.providers = [];
-
-		// Initialize posthog provider
-		posthog_provider.init();
-		if (posthog_provider.enabled) {
-			this.providers.push(posthog_provider);
-		}
 
 		// Initialize pulse provider
 		pulse_provider.init();
@@ -41,7 +33,7 @@ class TelemetryManager {
 
 	can_enable() {
 		let sentry_available = Boolean(frappe.boot.sentry_dsn);
-		return this.posthog_available || this.pulse_available || sentry_available;
+		return this.pulse_available || sentry_available;
 	}
 }
