+++ frappe/utils/telemetry/__init__.py
@@ -4,16 +4,11 @@ WARNING: Everything in this file should be treated "internal" and is subjected t
 removed without any warning.
 """
 
+from contextlib import suppress
+
 import frappe
 from frappe.utils import getdate
 from frappe.utils.caching import site_cache
-
-# posthog provider
-from .posthog import POSTHOG_HOST_FIELD, POSTHOG_PROJECT_FIELD
-from .posthog import capture as ph_capture
-from .posthog import capture_doc as _ph_capture_doc
-from .posthog import init_telemetry as _init_ph_telemetry
-from .posthog import is_enabled as is_posthog_enabled
 
 # pulse provider
 from .pulse.client import capture as pulse_capture
@@ -24,23 +19,26 @@ def add_bootinfo(bootinfo):
 	bootinfo.telemetry_site_age = site_age()
 	bootinfo.telemetry_provider = []
 
-	if is_posthog_enabled():
-		bootinfo.enable_telemetry = True
-		bootinfo.telemetry_provider.append("posthog")
-		bootinfo.posthog_host = frappe.conf.get(POSTHOG_HOST_FIELD)
-		bootinfo.posthog_project_id = frappe.conf.get(POSTHOG_PROJECT_FIELD)
-
 	if is_pulse_enabled():
 		bootinfo.enable_telemetry = True
 		bootinfo.telemetry_provider.append("pulse")
 
 
 def capture(event, app, **kwargs):
-	if is_posthog_enabled():
-		ph_capture(event, app, **kwargs)
-
 	if is_pulse_enabled():
 		pulse_capture(event, app=app, **kwargs)
+
+
+def capture_doc(doc, action):
+	with suppress(Exception):
+		age = site_age()
+		if not age or age > 15:
+			return
+
+		if doc.get("__islocal") or not doc.get("name"):
+			capture("document_created", "frappe", properties={"doctype": doc.doctype, "action": "Insert"})
+		else:
+			capture("document_modified", "frappe", properties={"doctype": doc.doctype, "action": action})
 
 
 @site_cache(ttl=60 * 60 * 12)
@@ -50,8 +48,3 @@ def site_age():
 		return (getdate() - getdate(est_creation)).days + 1
 	except Exception:
 		pass
-
-
-# for backward compatibility
-init_telemetry = _init_ph_telemetry
-capture_doc = _ph_capture_doc
