+++ pyproject.toml
@@ -84,7 +84,6 @@ dependencies = [
     "google-api-python-client~=2.188.0",
     "google-auth-oauthlib~=1.2.4",
     "google-auth~=2.48.0",
-    "posthog~=5.0.0",
     "vobject~=0.9.9",
     "pycountry~=24.6.1",
     "websockets~=15.0.1",
